Created routes folder, cleaned up server.js
