+++ server.js
@@ -6,7 +6,7 @@ const planetList = ["EARTH", "SATURN", "MARS", "URANUS", "VENUS", "NEPTUNE", "JU
 const api_key = process.env.api_key;
 const db = require('./models');
 require('dotenv').config();
-
+require("./routes/api-routes.js")(app);
 
 // Blast off!
 app.use(express.static('public'));
@@ -21,159 +21,8 @@ const exphbs = require('express-handlebars');
 app.engine('handlebars', exphbs({ defaultLayout: 'main' }));
 app.set('view engine', 'handlebars');
 
-// Integrate routes
-// const routes = require('/.js');
-// app.use(routes);
-
 // Allow server to grab any port, but set 7000 as standard default
 const PORT = process.env.PORT || 8080;
-
-function call_api(finishedAPI) {
-  request('https://api.nasa.gov/planetary/apod?api_key=0uKX4xKAlZRAVq4uqHlrUTbiHmQ7et6zW7QvbAi6&hd=false', { json: true }, (err, res, body) => {
-    if (err) { return console.log(err); }
-    if (res.statusCode === 200) {
-      finishedAPI(body);
-    };
-  });
-};
-
-
-//multiple API calls
-const fetchApiInfo = async (url) => {
-  console.log(`Fetching ${url}`)
-  const astroInfo = await axios(url)
-  return astroInfo;
-};
-
-const fetchAstroInfo = async (tickers) => {
-  // const requests = tickers.map(() => {
-  const url = `https://api.nasa.gov/planetary/apod?api_key=0uKX4xKAlZRAVq4uqHlrUTbiHmQ7et6zW7QvbAi6&hd=false`
-  return fetchApiInfo(url)
-    .then((res) => {
-      return res
-    })
-  // })
-  // return Promise.all(requests)
-};
-
-app.get('/astronomy', async function (req, res) {
-  const data = await fetchAstroInfo();
-  const results = {};
-  results.response = data;
-  console.log(results);
-  res.render('astronomy', results);
-  // }
-});
-
-// Create astro api/search
-app.post('/astronomy', async function (req, res) {
-  var sp = await req.body.astroSearch.split(",");
-  const data = await fetchAstroInfo(sp);
-  const results = {};
-  results.response = data;
-  console.log(results);
-  res.render('astronomy', results);
-  // }
-});
-
-
-//Solar system API
-
-function solarSystem_api(solarApi) {
-  request('https://api.le-systeme-solaire.net/rest/bodies', { json: true }, (err, res, body) => {
-    if (err) { return console.log(err); }
-    if (res.statusCode === 200) {
-      solarSystem_api(body);
-    };
-  });
-};
-
-//multiple API calls
-const fetchSolarSystemInfo = async (url) => {
-  console.log(`Fetching ${url}`)
-  const solarInfo = await axios(url)
-  return solarInfo;
-};
-
-const fetchSolarInfo = async (tickers) => {
-  // const requests = tickers.map(() => {
-  const url = `https://api.le-systeme-solaire.net/rest/bodies`
-  return fetchSolarSystemInfo(url)
-    .then((res) => {
-      return res
-    })
-    .catch(err => {
-      console.log(err);
-    }
-    )
-}
-
-
-app.get('/solar', async function (req, res) {
-  const data = await fetchSolarInfo();
-  const results = {};
-
-  //it would still help us GREATLY if we could figure out how to use url parameters instead of .filter to only get planets
-  results.bodies = data.data.bodies
-    .filter(body => body.isPlanet && planetList.includes(body.englishName.toUpperCase()))
-  // .map(body => addImageURL(body));
-
-  console.log(results.bodies);
-
-  res.render('solar', results);
-
-});
-
-//eventually puth this in its own file
-// function addImageURL(bodyObject) {
-//   let route;
-//   switch (bodyObject.englishName.toUpperCase()) {
-//     case "EARTH":
-//       route = "/images/earth.gif"
-//       break;
-//     case "MARS":
-//       route = "/images/mars.gif"
-//       break;
-//     case "SATURN":
-//       route = "/images/saturn.gif"
-//       break;
-//     case "URANUS":
-//       route = "/images/uranus.gif"
-//       break;
-//     case "NEPTUNE":
-//       route = "/images/neptune.jpg"
-//       break;
-//     case "JUPITER":
-//       route = "/images/jupiter.gif"
-//       break;
-//     case "VENUS":
-//       route = "/images/venus.gif"
-//       break;
-//     case "MERCURY":
-//       route = "/images/mercury.gif"
-//       break;
-//   }
-
-//   bodyObject.image_url = route;
-
-//   return bodyObject;
-// }
-
-
-
-
-
-
-// Create solar api/search
-app.post('/solar', async function (req, res) {
-  var sp = await req.body.solarSearch.split(",");
-  const data = await fetchSolarInfo(sp);
-  const results = {};
-  results.response = data;
-  console.log(results);
-  res.render('solar', results);
-  // }
-});
 
 // Start Server
 db.sequelize.sync({ force: true }).then(function () {
